Fixes cache decorator for python 3.7

+++ run_tests.py
@@ -15,10 +15,8 @@ import sys
 import argparse
 import importlib
 import unittest
-import coverage
 import cProfile
 import logging
-import pycodestyle
 import io
 import contextlib
 import subprocess
@@ -27,14 +25,23 @@ from timeit import default_timer as timer
 parser = argparse.ArgumentParser()
 parser.add_argument('--max-test-level', type=int, default=1)
 parser.add_argument('--test-levels', type=int, nargs='*', default=None)
+parser.add_argument('--coverage', dest='coverage', action='store_true', default=True)
+parser.add_argument('--no-coverage', dest='coverage', action='store_false')
+parser.add_argument('--codestyle', dest='codestyle', action='store_true', default=True)
+parser.add_argument('--no-codestyle', dest='codestyle', action='store_false')
+
 args = parser.parse_args()
 if args.test_levels is None:
     args.test_levels = list(range(args.max_test_level+1))
+if args.coverage:
+    import coverage
+    os.system('rm -f .coverage')
+if args.codestyle:
+    import pycodestyle
+    os.system('rm -f .codestyle')
 
 # Clean up:
-os.system('rm -f .coverage')
 os.system('rm -f .profile')
-os.system('rm -f .codestyle')
 os.system('rm -rf htmlcov')
 
 # If only clean required, exit here:
@@ -42,14 +49,15 @@ if len(sys.argv) > 1 and sys.argv[1] == 'clean':
     sys.exit(0)
 
 # Produce the codestyle report:
-style = pycodestyle.StyleGuide(paths=['vayesta/'], report=pycodestyle.FileReport)
-f = io.StringIO()
-with contextlib.redirect_stdout(f):
-    report = style.check_files()
-with open('.codestyle', 'w') as file:
-    file.write('\n'.join(report.get_statistics()))
-    file.write('\n\n')
-    file.write(f.getvalue())
+if args.codestyle:
+    style = pycodestyle.StyleGuide(paths=['vayesta/'], report=pycodestyle.FileReport)
+    f = io.StringIO()
+    with contextlib.redirect_stdout(f):
+        report = style.check_files()
+    with open('.codestyle', 'w') as file:
+        file.write('\n'.join(report.get_statistics()))
+        file.write('\n\n')
+        file.write(f.getvalue())
 
 # Get a list of untracked files to ignore:
 untracked = subprocess.run(['git', 'ls-files', '--others', '--exclude-standard', 'vayesta/'], capture_output=True, text=True).stdout
@@ -63,10 +71,11 @@ untracked += [
 ]
 
 # Start the coverage report:
-print("Coverage omitting files:")
-[print(f) for f in untracked]
-cov = coverage.Coverage(omit=untracked)
-cov.start()
+if args.coverage:
+    print("Coverage omitting files:")
+    [print(f) for f in untracked]
+    cov = coverage.Coverage(omit=untracked)
+    cov.start()
 
 # Import Vayesta (inside coverage scope) and restrict logging to WARNING:
 import vayesta
@@ -103,9 +112,10 @@ print("Finished %3d tests in %.0f s" %  (ncount_tot, timer()-t0))
 prof.disable()
 prof.dump_stats('.profile')
 
-# End the coverage report:
-cov.stop()
-cov.save()
+if args.coverage:
+    # End the coverage report:
+    cov.stop()
+    cov.save()
 
-# Generate HTML report for coverage:
-cov.html_report()
+    # Generate HTML report for coverage:
+    cov.html_report()
